result type of formatSize + (docID + docStoreOffset) * 8) is int32. When docID is max then 268435455, the result wiil overflow of int32 and get negative start offset

+++ src/core/index/FieldsReader.cpp
@@ -148,7 +148,7 @@ int32_t FieldsReader::size() {
 }
 
 void FieldsReader::seekIndex(int32_t docID) {
-    indexStream->seek(formatSize + (docID + docStoreOffset) * 8);
+    indexStream->seek(formatSize + (docID + docStoreOffset) * (int64_t)8);
 }
 
 bool FieldsReader::canReadRawDocs() {
